fix: use correct ticket name in redirect form
The current code does not work, at least in my environment, and
has not worked for a long time, I'm guessing. TouchNet staff let
me know the error on their end, and it was that the ticket name
initially sent (which includes a date stamp) was not the ticket
name eventually used (which was just the user id without a date
stamp). This fixes the issue.