chore: remove request as a direct dependency
The dependency request has been deprecated and unmaintained
for over 5 years. It is only used so it can be wrapped in
a utility function to allow promise-based requests. This is
the role of fetch() in JavaScript which was not available
in Node.js for many years but has been for some time now.

This removes `request` and the wrapper `requestp()` and
replaces them with `fetch()`. Some additional error-checking
is performed.

The request module will still appear in the dependencies
because it is used by almarestapi-lib. That will be
addressed in a subsequent pull request.