chore: remove request as a secondary dependency
The dependency request has been deprecated and unmaintained
for over 5 years. It is used by almarestapi-lib which itself
appears to likely be unmaintained, with no updates in 3 years.
The connector only uses fraction of the functionality of
almarestapi-lib, all of which is compact and maintainable once
switched to fetch() from the obsolete request().